Backend/pyproject.toml
[project]
name = "backend"
version = "0.0.1"
description = "Testify Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.2",
    "greenlet>=3.2.4",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.14.5",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.27",
    "unidecode>=1.4.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = ["alembic/**"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short", 
    "--strict-markers",
    "--color=yes"
]
